requirements.txt
python-telegram-bot==20.3
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7
requests==2.31.0